Periodically index and download movies (#20)
* pull some methods out of the manager.go file

* start at running things periodically
